// src/app/shared/types/api-types/bire-variant-version-dto.interface.ts
export interface BireVariantVersionDTO {
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  variantName?: string;
  variantRevisionNumber?: string;
  variantRevisionDate?: Date;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  familyFunction?: string;
  familyName?: string;
}